Extract SnapshotVersion to agave-snapshots
