+++ snapshots/src/lib.rs
@@ -2,5 +2,8 @@
 
 mod archive_format;
 mod snapshot_interval;
+mod snapshot_version;
 
-pub use {archive_format::*, snapshot_interval::SnapshotInterval};
+pub use {
+    archive_format::*, snapshot_interval::SnapshotInterval, snapshot_version::SnapshotVersion,
+};
